requirements.txt
python-dotenv
requests
elasticsearch
mysql-connector-python
alembic
Faker
pymysql